Fix not handling notifications (#1)
The behaviour when we were receiving a push notification was not being handled. This logic is in the SDK so the plugin should not override this logic assigning itself as the delegate of the UNUserNotificationCenter.
We added also the possibility to print the firebase token so we can send custom push notifications to individual users.